requirements.txt
docker
flask
flask_swagger_ui
flask-restful
typer
kivy
kivymd
python-dotenv
requests
typing_extensions
woocommerce
opencv-python
